+++ libca.em/src/em_eimui/xmake.lua
@@ -1,1 +1,0 @@
--- test target moved to libca.em/unittests/em_eimui/xmake.lua