+++ libca.em/src/em_ota/xmake.lua
@@ -1,6 +1,0 @@
-target("libca.em_ota")
-    set_kind("static")
-    set_group("em")
-    add_files("partition.c")
-    add_deps("libca.em_base")
-    add_includedirs("..", {public = true})