+++ libca.em/src/em_util/doubly_list.c
@@ -1,2 +1,0 @@
-#include "doubly_list.h"
-